requirements.txt

fastapi==0.122.0
grad_cam==1.5.5
hnswlib==0.8.0
httpx==0.28.1
matplotlib==3.10.7
numpy==2.3.5
opencv_python==4.12.0.88
opencv_python_headless==4.12.0.88
Pillow==12.0.0
python-dotenv==1.2.1
torch==2.9.0
transformers==4.57.1
ultralytics==8.3.220
uvicorn==0.38.0
